docker/requirements.txt
# SPDX-FileCopyrightText: 2025 Idiap Research Institute <contact@idiap.ch>
# SPDX-FileContributor: Amir Mohammadi  <amir.mohammadi@idiap.ch>
#
# SPDX-License-Identifier: MIT

# server dependencies
fastapi[standard]==0.115.12
python-multipart==0.0.20
numpy==1.26.4
pillow==11.0.0

# trufor dependencies
# torch==2.6.0
# timm==1.0.15

# cnet dependencies
torch==1.13.0
torchvision==0.14.0
albumentations
timm